butch-core: auto-add deps.stage[0-9] to DEPS when building a stage
building a stage is a somewhat extraordinary situation:
for bootstrapping purposes it needs to enforce a specific order of
packages and specific compiler versions to get to the point where we
have a fully equipped system.
these strict enforced orders usually get into the way outside of
bootstrapping scenarios (regular builds) and appear a bit arbitrary.
we solved this issue elegantly for stage0 by putting its deps into
a special deps.stage0 section.
this does make sense as well for stage1, and probably also for other,
not-yet-invented stages.
so we add it as a general behaviour that when such a package name is
passed, an explicitly dep on deps.stageX is added.