firefox: make build reproducible
the major issue was a `BuildID`, that went into almost all binaries
and several config files.

for example in application.ini:
BuildID=20170306191502

the build id gets passed along in the build system as MOZ_APP_BUILDID,
and it's derived from MOZ_BUILD_DATE.

the .chk files are some "signatures with ephemeral private keys that
are generated at signature time" which are only needed with FIPS and
thus can be safely removed.