add a generic workaround for packages using intltool insisting on perl
basically, it's the fault of intltool.m4 [0]. even when --disable-nls was
specified, it will still check for existence of 3 tools from the intltool
suite, plus for perl and it's XML::Parser module. so any package using
intltool.m4 in their autoconf scripts will pull this dependency into
configure, even though it's later not needed during the entire build process.
with one exception: .desktop files will get processed once with intltool-merge.

this here is a script that emulates this one particular usage using sed,
and otherwise just gives the answers intltool.m4 expects them to give,
or fails with error on any other usage. it only contains a fake perl that just
gives satisfying answers to the 3 configure tests.

the second script is a companion that can be sourced from package recipes:
it automatically copies the intltool fake script into a subdir, creates
symlinks for all the tools and exports the required environment variables.

[0]: https://bugs.launchpad.net/intltool/+bug/904647