intltool: also provide at least a basic XML changing facility
this functionality is used unconditionally by shared-mime-info and gimp,
even with --disable-nls.

the task is to change <_name>foo</_name> style tags into <name>foo</name> [0].

our regex is not perfect yet, it may also need to handle the case of tags like
<_tag foo=bar>baz</_tag>. in order to support that we need to check for a
closing `>` or whitespace, and replicate it with a second back reference.
for the moment though it looks sufficient as-is.

[0]: https://davesteele.github.io/debian/development/2015/12/10/i18n-build-flow/