alsa-lib: add patch to avoid unnecessary forking
wordexp() is implemented in a way that it forks and executes the
shell. there were reports in the musl irc channel that this caused
considerable lag with some applications. this patch was also posted
there, and it makes sense that when a path is absolute, wordexp
doesn't need to be called.