python: properly setup LC_CTYPE locale on startup
any program dealing with text should setup setlocale(LC_CTYPE, "")
on startup. python doesn't do so (well it does once, to get
nl_langinfo(CODESET), but then sets it back to the previously stored
value). this caused issues with libedit (readline replacement) which
due to this omission refused to accept UTF-8 input, which in turn
caused the python REPL to shutdown due to receipt of (according to
its own interpretation) EOF when entering any UTF-8 character.

while debugging this issue i encountered special-case code in the
readline module which deals with libedit specific quirks by detecting
it at runtime using a version string (whether it is active can be
queried by printing readline.__doc__ in the REPL).

this code though is only enabled if __APPLE__ is set so i tried to
enable it for our use case too.
however, the resulting readline module behaves buggy, for example
tab completion doesn't work.
due to this, i commented the necessary changes out again.

the tab-complete issue in libedit mode might be caused by a combination
of our old configure script patching which allowed to compile against
libedit without knowlegde of this compatibility code.
maybe it's interesting to look at this in the future again if any
issues with the current readline modules pop up, so the commented out
code is left for the record.

if any issues due to our new UTF-8 patch occur (unlikely), an alternative
would be to patch the readline module to call setlocale(LC_CTYPE, "")
before the readline library is initialized (it seems that libedit,
unlike readline, only queries the codeset on startup, not on every library
call - because they do actually save the current locale, and set
setlocale(LC_CTYPE, "") just before any readline api call - just not on
initialization...)

on a side note, it seems the python developers are *kinda* aware of the
issue, however they're working on an over-engineered solution:
https://www.python.org/dev/peps/pep-0538/
it seems that because of broken android behaviour, their plan to get a
proper UTF-8 environment by default is basically:
 - setenv("LC_ALL", "C.UTF-8", 1);
 - setenv("LANG", "C.UTF-8", 1);
 - setlocale(LC_ALL, "");
which actually would override the user's configured locale in all child
processes.
the proper fix is actually just doing what this patch does,
plus something like an ifdef hack for the broken android behaviour cited
in that PEP.