fix kernel build with latest busybox
with latest busybox, apparently a bug in sh got fixed that makes

    set -e ; do_some_stuff || more_stuff

work as intended, i.e. an error is thrown if do_some_stuff failed.
this causes a build failure in the kernel which tries to generate
timeconst.h using a bc script, which uses GNU-specific features
that are not supported by plan9's bc, which we currently use.
this step failed in the past too but went unnoticed since the
`set -e` didn't kick in as intended, so the produced timeconst.h
was empty, but apparently there's fallback code in the kernel that
this didn't matter.

there are several possible workarounds, one would be to just remove
the set -e, or patch the bc call to just return true, to install
a GNU-compatible bc program, or, what we opted for, to reinstantiate
landley's linux-noperl-timeconst.patch.
timeconst.h was previously generated by a perl script, and landley
wrote an implementation in C to replace it, but the kernel folks
instead adopted a GNU-bc specific bc script to generate the values.

patch from https://raw.githubusercontent.com/landley/aboriginal/master/sources/patches/linux-noperl-timeconst.patch

the part deleting the bc script was removed, since it's not
necessary and it makes the patch fail whenever something in the
bc file was changed.