update gcc424, support C++, and crosscompilation
this is a huge hackjob, but it's now possible to both crosscompile
gcc424 (but when including C++ support, only using the musl-cross
4.2.1 cross-compiler, other newer crosscompiler use a different subset
of C++ which makes the build fail in one way or another).
since butch's optional feature support is not yet finished, one currently
needs to "butch download g++424" first, and then use
"option_cplusplus=1 butch rebuild gcc424" to build gcc424 with C++.

further improvements are backports of the support for sub-arch dynlinker
names like ld-musl-mipselsf.so.1, though (hack hack) this support is only
given for *native* builds which use the fake GNU triple, since for cross
builds (unless a glibc crosscompiler is used) the musl triple will kick in
and will cause the musl support patches to do their magic (which does not
support the added dynlinker names at this point). if the GNU triple
would be used, our sed magic would cause the right dynlinkers to be used.

another backport is support for compiling it with newer GCC > 5.

one major remaining issue is that if C++ support is requested, only
libstdc++.a is built, but not the shared library. this is good enough
for stage0 support, but not really for a proper stage1 toolchain.

but then, as this GCC basically only exists for the sake of bootstrapping
ARM natively, this is OK for the time being.
note that this compiler supports only armv6 and max vfp fpu, so its not
really a toolchain one wants to keep anyway, even though it is probably
much faster and lighter than later GCC versions.