build-stage: revert to building only C toolchain
since commit 7620293531ab71fd03e02cd5aff720273cba5f7c we changed
stage0 to build gcc3 with C++ support, so we could support a stage1
gcc newer than 4.7.4, which is last gcc written in C.

that however opened a can of worms, since building a C++ toolchain
is much more complicated than a C toolchain. we encountered a lot
of issues with incompatibilities of the toolchains provided by
the different distributions. the major issue was that recent binutils
emit a new relocation type which our binutils version can't deal
with, and there's no way to turn that off.

so now, since we already have gcc 6.3.0 as our stage1 compiler, and
thus require a C++ toolchain, we opt for building gcc3 twice:

in the foreign environment of the bootstrap distribution, only a
minimal C compiler is built, which is then used at the very beginning
of stage1 to build gcc3 again, this time with C++ support. since
we control the environment we get in stage0, this is safe.
that C++ toolchain is then used to build the final stage1 compiler.

in order to avoid duplication of the gcc3 package, we introduce
a novel concept in sabotage: both stage0-gcc3 and stage1-gcc3 are
symlinks to gcc3. the same could and should be applied to stage0-musl
and musl too.

while we were at it, the gcc3 package was modernized, and the gcc424
package which is used to bootstrap arm platforms natively (since
gcc3 doesn't support arm eabi yet) was adapted too. it is arguable
whether we should default non-arm bootstrap to gcc424 too, but since
it is slightly slower to build, the only benefit would be unified
handling of the bootstrap toolchains.

this commit is based on a patch by @JillTyrell.