lesstif2: enable xrender/xft support
the configure code to check for xrender is broken, and so is the code
adding the libraries when the detection code has been worked around...
possibly because of those bugs, the xft code paths have never been tested.
i have yet to see an advantage of activating the xft support, but committing
this for documentation purposes, since xft and xrender have to be installed
anyway due to other dependencies.