libtool.cross: do not fail when no "tagged configuration" was infered
our libtool by default has a tagged configuration (whatever the purpose of
that) for C++, without which no C++ programs can be linked.
if no tags are given in available_tags, libtool doesn't try to use any,
however under some circumstances, the function is called and when the C++
tag doesn't match, it bailed out with an error.
now we supress erroring out, while adding a second configuration that resets
the CC variable to the C compiler, if the CXX tag didn't satisfy libtool.

this fixes build with sdl2-image, and probably also with other packages where
this issue was experienced in the past, such as recent versions of php.