add preliminary support for building packages with ccache
to use it, basically you need to install and configure ccache, and
then build selected (!) packages with

    USE_CCACHE=1 butch rebuild $pkg

since most packages build very pick, using ccache globally doesn't make
sense. also, the cache needs to be restricted to use only a certain amount
of HDD space, and when building everything with it this would quickly
overflow.

so my recommendation is to use it only for certain packages that take very
long to build, but are often upgraded. best examples are kernel and firefox.