butch template: add workaround for gnu tar idiocy
since the host environment in stage0 is out of our control, gnu tar
may be used, which has an odd behaviour: when the user is root, files
extracted get owned by the UID that is stored in the tarball, even when
that UID doesn't exist on the host system.
this causes issues later on when we want to rebuild a package
(example: gnu make) inside the super chroot, which then will fail due to
permission errors (even when we were root outside of the chroot).
this is due to how super_chroot works.