busybox: fix dangerous nonsensical omission to create links
if applets were found in the rootfs' /bin, they were not added as
symlinks, the original intention was probably to prevent busybox
rebuilds from overwriting replacements like wget, but that didn't
work anyway because no_unlink packages did never clean up their
staging directories in /opt (this will change in the next commit).
apart from that, it's a very nasty behaviour to change what's
getting into a package based upon what's installed on the system.
this caused a nasty surprise and breakage when testing the new
code that intends to clean up staging dirs of no_unlink packages.