remove digimend
this package does things way too hackish, like assuming there is a stale
kernel build dir, and then mess around inside that other package's build
dir. if the original author is interested in getting this fixed, i'd suggest
to look at how the `perf` package does things, and make according changes,
also properly document deps, then we can re-add it.
as it is now, this package is best kept in a private fork.