remove flatpak
this package doesn't compile on standard sabotage, it would either
require building glib with dbus and python support, so the `gdbus-codegen`
tool is provided, which is not really desirable, or providing fake files
that can be linked in instead of the gdbus-codegen-produced ones.
as such, this package should be kept on private repo until it's fixed for
general consumption.