gcc640: initial java support
patch taken from alpine.
currently in order to build, the 2 packages in the deps.java section
need to be manually installed, then gcc needs to be rebuild with
option_java=1 environment variable set.
the resulting GCC is not capable of building GNU classpath (raises
internal compiler error), but this may not be needed to bootstrap
openjdk. using ecj-4.5.jar instead of 4.9 was already tried, but
doesn't fix the classpath issue.

most of the work is based on 677133150639aa5bb0018f0eed70e1d3d4989220.