default kernel to new 4.14 LTS branch (4.14.44)
CONFIG_UNWINDER_ORC is a new "trick" the kernel uses to create
non-standard ELF files, which are not compatible with GNU binutils.
fortunately it can be disabled.

CONFIG_USB_PCI is a new option that allows to remove a couple bytes
for embedded boards that don't have USB/PCI bridges, but having the
option off (default) breaks USB on ordinary PC hardware.