linux.config: enable IWLWIFI driver
this driver has an odd constraint: if the driver is statically linked
into the kernel, the firmware must be linked into the kernel too [0]!
this forces us to either make the kernel package dependent on the linux-
firmware package (so the firmware can be found at compile-time) for
static linking, or build that kernel module dynamically, unlike all
others.
i opted for the second option, since it's less restrictive and the kernel
package is already prepared for dynamic modules anyway.

[0]: https://wiki.gentoo.org/wiki/Iwlwifi