kernel: fix kernel module install
in order to install the modules into the right path, we must infer
the final linux version from the last applied patch, rather than
using the base version of the main linux tarball.

note that the depmod/modprobe program of the "kmod" package create/use
binary versions of modules.* with the suffix .bin; while busybox modprobe
works with plaintext versions.