xf86-video-intel: bump to git version of 20170325
the latest official release is from 2014, so we're forced to use
git version to get a somewhat up-to-date driver.

since the patch (as made by alpine linux) is > 1MB, we download it
rather than embedding it into KEEP.

what leaves a bitter taste is that due to several changes in the
build system, we're forced to regenerate the autoconf files and thus
pull in dependencies to autoconf and automake, unless we provide
the pre-made configure script ourselves.
it'd be preferable if the developers could be bothered to make a
release.

this update fixes several issues with UHD Graphics 620 chipset, such
as tearing effect in GTK+2 apps, and failure to fullscreen mpv with
the xv video out (-vo=xv) which is orders of magnitude faster than
the GL one.