bison: fix cross- and native compile on aarch64
updating config.sub fixes detection of the aarch64-*-* triplet from
`$CC -dumpmachine` for crosscompilation.

for native compilation, where usually no triplet is passed to configure,
configure tries to "guess" which arch it is running on via `config.guess`,
which is pre-aarch64 as well (i.e. too old), and then fails.

since config.guess is a really big file, we want to avoid checking it into
KEEP/, and so we specify the toolchain triplet manually via --build, which
prevents config.guess from being executed.

this should only be necessary for very few packages, that have ancient copies
of config.guess in their tree.