obconf: remove failing patch hunks
patch provided by @slackmlinux

bg info: originally we didn't have set -e in the build scripts.
that's why nobody noticed the patch hunks that didnt apply.
then we added it and a contributor stumbled upon that error.
and rather than fixing it, he decided to just print a warning instead.