gcc3: do not build in crosscompile setup
this package is pulled in via stage1-gcc3 which is in gcc650 (stage1
default compiler)'s [deps.stage1] section.
when we cross-compile we don't want to build it though, so we would
actually need a combination of deps.stage1 and deps.host, to properly
articulate the dependency without hacks like this.