python: fix some cross-compile related issues (host-only)
when used as the host python install, this fixes some issues with cross-compile
of python packages. in a cross-compile scenario, python now uses the pkg-config
sysroot environment variable to fix up -I and -L paths to point into the sysroot
rather than the host installation.

additionally:

1) do not add gratuitous -L parameters to CFLAGS
2) use OPT parameter for CFLAGS, as this will later be used as base CFLAGS
   if CFLAGS are passed in the environment.
3) empty the OPT parameters saved at python build time.

this means that if no CFLAGS are exported, the same CFLAGS used for the build
are reused, otherwise only those explicitly passed.

this fixes build errors on non-x86 targeting crosscompiles with -mtune=generic
leaking from the host and breaking the build as this is an x86 specific flag.

note that packages using setup.py install need a prefix containing
$butch_install_dir$butch_prefix passed to get their installation target
straight.
this will be tacked in follow up commits where necessary.
according to https://docs.python.org/2/install/index.html#alternate-installation-unix-the-prefix-scheme
it seems there should be no other sideeffects when using --prefix to fulfill
the purpose of $DESTDIR and $PREFIX in one.