add palemoon browser
this version is from december 2018, when work on getting it compiled started.
adding a newer version will be attempted soon. functionally, this release seems
to be identical to our latest firefox release firefox 52.0.9esr, as it exhibits
the same problems with latest jquery code.
interestingly, even though the tarball is labelled palemoon, an executable named
`basilisk` is produced, which seems to be the name of a second browser project
by the same people. maybe this was a mistake that only happened in this specific
release.