add gpc, the GNU pascal compiler
this is mostly based on the gcc3 recipe, but the number of necessary
changes has been intrusive enough that i opted for putting this into
its own package rather than adding yet more conditionals to the gcc3
package. since gpc builds the C compiler as part of the build process
anyway, and it is only 6MB, i opted for installing it as well, as a
bonus.

thanks go to mirabilos from MirBSD, who has been very helpful getting
this accomplished. it seems MirBSD is one of the last OS/distros to
still ship GPC. 2 of the patches are taken from MirBSD, fixing crucial
build errors. the cumulative patch from MirBSD is temporarily stored
on http://ftp.barfooze.de/pub/sabotage/tmp/mirbsd-gpc-gnu-pascal-20060325.patch
as a reference in case further issues are encountered.