butch printsec: add %include (recipe)% preprocessor directive
using this, build sections having lines like
%include pkgname%
on a line of its own can now include the build section contents of another
package. this is meant as a mechanism to e.g. streamline all gcc* packages
into a single master package, which can then select the set of patches to
apply via the package name, or some options set via var section or before
the include directive. it also allows to create packages that have different
dependencies, but (almost) the same recipe, such as e.g. a gcc-gcj package,
which would add java related dependencies and toggle an option to build
java on in the shared recipe body.