README: 'world' -> 'desktop'; removed old sections.
-- THANKS
The THANKS section came from a file included in the first hundred
or so commits of the original Sabotage repository[0]. A lot has
changed in the following 6,000~ commits, but the THANKS section has
not. It has been a point of contention in several situations and
it seems it would be simplest to drop it.

We still wish to extend our thanks to leahneukirchen, dalias,
niklata, garbeam, pikhq, xmw, gaf and Arch Linux!

[0]: https://github.com/leahneukirchen/sabotage

-- CONTACT
While the old mailing list is a historical document, any value it
has should be documented in the COOKBOOK and Wiki instead.