libx11: fix x-compile on non-sabotage hosts
the configure script runs `pkg-config --variable=includedir xproto` to look
for xproto header files, and for some reason both pkg-config and pkgconf ignore
PKG_CONFIG_SYSROOT_DIR for this purpose (but not PKG_CONFIG_LIBDIR), so the pkg-
config file from the sysroot is used to return "/include", which doesn't exist
on a typical linux host.
this also used to work only by accident in the past on sabotage systems (via
them having the xproto headers installed on the host fs).