config.cross: improve user-friendliness
it's quite easy to mess up the path of the toolchain, in which case
butch use will fail, or the rootfs base R, which can result in quite unexpected
things when running setup-rootfs, in both cases requiring to debug the paths.
also default to musl-cross-make base path, instead of musl-cross.