kernel*: use cumulative patch against base tarball
this removes the need to have one package for each patchlevel.
this speeds up the build and clutter in pkg/ and KEEP/, but on the
negative side the user has to download a new kernel base tarball, although
only once, as the cumulative patch kernel.org supplies applies to the .0
revision.