quickjs: fix crosscompile
the crosscompiler variable was stuffed into a -D preprocessor string argument
to qjsc.c with only basic quoting, and since our CC variable contains spaces
this broke.
this is not only wrong from the quoting aspect, but also, assuming
a cross-compiled binary is run on the targetted architecture, this would then
result in qjsc trying to run the crosscompiler on the target which doesn't make
sense.