xz: explicitly enable symbol versions
the update to 5.2.5 caused programs linked to xz or one of the libs
using it, such as libxml2 or libarchive, to fail at runtime due to
failure to find versioned symbols @XZ_5.0 like this:

/lib/libarchive.so: undefined reference to `lzma_end@XZ_5.0'

underlying cause was that 5.2.5 updated its config.guess, which now
detects musl via an ldd check, and the symbol versioning code gets
only enabled if gnu is detected.

arguably it would be better to have symbol versioning turned off,
but since we didn't notice this in the past its more convenient
to just keep compiled apps working rather than forcing everyone
to rebuild all users of xz, as would be required for a soname bump.