update all kernels to latest patchlevel
when we last bumped kernel patches, a new breakage was introduced in the
revision just following the one we checked in, e.g. we added 4.14.192
while 4.14.193 introduced the breakage, and so on. at that time we refrained
from checking in the very latest broken patchlevel in the hope this would be
fixed until the next time we bump revisions, but that didn't happen.
so the patch fixing the issue (which was merged to mainline kernel in may)
has been added now, it will only be applied if the header introducing the
issue contains the keyword offsetof.