sdl2: bump to 2.0.14
unfortunately the configure script is broken with dash/busybox sh[0].
the hacks/bash dependency need to stay until the issue is fixed in
a release.

when running configure with bash, the resulting libtool uses bashisms
which then fail when running make, therefore we use our libtool.cross
until the bash dependency goes away.

[0]: https://bugzilla.libsdl.org/show_bug.cgi?id=4261