palemoon: bump to 28.17.0
our patch for high cpu usage has made it into upstream, however conditionally:
the about:config option "network.websocket.timeout.clamped" which needs to be
set to true.
the patch was later reverted so it's possible that option won't work anymore
in the next release, but apparently a better fix was found.

https://repo.palemoon.org/MonchildProductions/UXP/issues/1695
https://repo.palemoon.org/MoonchildProductions/UXP/commit/2fa993b5639e04c7e1d7403ecf9175a223ce50b4

the commit above is also interesting in that it shows how an about:config
item can be added and how it interacts with the C++ code.