gdb7*: fix "Signal ?" error when debugging certain threaded code
gdb handling of unknown signals is broken and causes those signals to not be
delivered to the target, which makes it hang and effectively makes debugging
the affected program impossible.
musl uses some internal signals which gdb doesn't know for setrlimit() impl.

with this patch, the problem is circumvented but not fixed.
after applying it one can handle them via

handle SIG34 nostop noprint

bg info:
https://sourceware.org/bugzilla/show_bug.cgi?id=23616