butch unpack: fix bad things from happening when unpacking busybox
with butch unpack "busybox_shiny_new_version" bad things happened:
namely busybox was unlinked and deleted and the whole unix command
line environment unavailable.

to fix that disaster i used nano to create a file called main.c
with the contents "int main() {}", gcc to compile that to a file
named busybox (gcc main.c -o busybox) so we have an executable file
of the right name, then curl otherpc:8000/bin/busybox > busybox.
that way the resulting busybox binary keeps the +x permissions
of the C compiler output, and finally
./busybox --install -s /opt/busybox/bin to make the applets available
again.

the butch-unpack fix is to simply check whether we deal with a no_unlink
package and then not rm it before unpacking.