glade: fix build without intltool/perl5-XML-Parser installed
it was only used to generate translations for the funky .desktop file.
apparently that went wrong in the last release (so we had to produce the
file manually), so the maintainers decided to put intltool inside the source
tree this time and override the location from where it's used.
a comment in configure says "Use the tools built into the package, not the ones that are installed".