butch_download_template: remove curl cookie-jar option
this switch, added in 7569032b510ddaa79e3ba400dbffb0b78a7f8dfd causes the
character device /dev/null to be deleted and recreated as an ordinary file,
when executed as root.

it's unknown whether the issue mentioned in the above commit message is fixed,
but time will tell - a failing download is in any case preferable over a
trashed /dev/null which will break almost any shell script in existence.

https://github.com/curl/curl/issues/6882