libressl: pretend getentropy() doesn't exist
if libressl detects that arc4random isn't available, but getentropy() is,
it will *only* use the latter to try to gather entropy.
getentropy() in turns calls getrandom(), which is a syscall introduced in
linux 3.17. if that fails (e.g. older kernel than that) libressl does a
raise(SIGKILL) to bring down the calling process without leaving a core
file which could contain precious secrets.

otoh, if getentropy() isn't available, getentropy_linux.c will be compiled
which first tries getrandom(), followed by opening /dev/urandom and so on.

therefore, disabling detection of getentropy results in no regression for
recent kernels with getrandom() - but additionally provides a proper fall-
back for older kernels.