sdl2-sound: switch to newer version that has proper midi playback
my patch was actually written against this very version, however it used
a tag of the now defunct hg repo, so i had to crunch through the git commit
list to find the corresponding commit of the new github repo.

according to the author, sdl-sound maintainers are currently preparing a
release, so hopefully we will soon have a proper tag to work with.
the release, once available, will replace libmodplug's midi player
with an integrated timidity version; it remains to be seen whether that's
an advantage.