add preliminary support for LTS kernel 5.10 as "kernel510"
caveat emptor: the kernel builds but hasn't been boot-tested yet, so
it's possible some CONFIG items need to be added to make it usable.

the regdb patch required 2 modifications:
- the hunk patching the documentation has been removed as the kernel
  documentation has been renamed from *.txt to *.rst
- recent kbuild doesn't support the ---help--- syntax anymore, so
  it was modified to just help, like the other entries surrounding
  it were already using.

the kernel build system now uses a tool called objtool, which despite
executing a subcommand called `check` which suggests the operation is
read-only, actually *rewrites* object files and adds some sections to
them that our binutils or elfutils can't deal with, while linking the
vmlinux binary the following error is thrown:

arch/x86/built-in.a: member arch/x86/events/core.o in archive is not an object

making objtool into a no-op fixes that, however this only works as long
as CONFIG_UNWINDER_ORC is not requested (which we don't).
according to documentation in objtool directory, objtool can be disabled
with the config option CONFIG_STACK_VALIDATION, however that turned out
not to be the case; the option can't even be toggled on or off in
`make menuconfig`.

a related bug report from gentoo can be found here:
https://bugs.gentoo.org/671760
gentoo traced it to a bug in elfutils, though i suspect in our case linux
developers assumed features only supported in bleeding edge binutils.

there are 2 further issues during the build:
kernel buildsystem uses "diff -I" which in GNU diff ignores lines based
on a regex, which busybox diff does not support. however this causes no
build error and replacing diff with diffutils' version doesn't result in
any differences in the resulting kernel binaries, as checked via
`butch checksum`. so it appears this issue can be safely ignored.

the other issue is that usr/gen_initramfs.sh uses "find -printf" which
again is a GNU-specific option to find which busybox does not support.
this does not result in a broken build either, and i suspect this issue
can be ignored as well as we don't use an initramfs provided by the kernel
build.
due to lack of a find implementation supporting the -printf flag however,
i could not verify that this has no effects on the resulting binaries.