fix stage0 build issues on host distros with compressed debug sections
distros using bleeding edge binutils default to produce compressed debug
sections, even if debug info was disabled via -g0 in CFLAGS, which leads
to error message like the following as soon as stage0 is built and the
user tries to use the toolchain (e.g., for building stage1):

/lib/gcc/x86_64-unknown-linux-gnu/3.4.6/../../../crti.o:
file not recognized: File format not recognized

or

/bin/ld: /lib/gcc/x86_64-unknown-linux-gnu/3.4.6/../../../crtn.o:
unable to initialize decompress status for section .debug_aranges

more info in this PR: https://github.com/richfelker/musl-cross-make/pull/73

other distros would probably just update to latest binutils to mitigate such
an issue, but this opens pandora's box and leads to a chain of consequences
such that gdb needs to be updated too, newer versions of which now use C++,
etc.
seen from a different PoV, sticking with an older and well-test binutils
version for the time being has the nice effect that binaries produced on
sabotage are way more portable and can be debugged with e.g. a 10 year old
gdb without issues.