util-linux: prevent install of libs, install most "applets"
someone compiling this package probably encountered a deficiency in
a specific busybox tool, so let's install almost all of the tools by default.
this also allows us to get rid of the partx and eject package.
otoh enabling more applets also requires to enable the libraries they depend
on, but we do not want them installed so they don't interfere with
libblkid/libuuid packages (the latter is provided by e2fsprogs).

since util-linux installs a lot of programs that replace busybox applets,
the recommended way to use it is to build it quickly after stage1 (so other
tools like e.g. wget, which too replace busybox applets, aren't installed
yet), and then `butch unlink` it and `butch relink busybox`, and use the
util-linux tools only from /opt/util-linux/bin, or link them manually to /bin.
alternatively it may be compiled in a cross-compile rootfs and the specific
binaries one is interested in taken from there, to avoid the unlink and relink
dance.