palemoon: bump to 30.0.0 (WARNING, see commit msg)
you're advised to wait with an update until the necessary addons such
as github-wc-polyfill, palefill, and others are updates to support it,
because palemoon 30 has a new profile format so you can't go back
to your old profile.

you might also want to test it as a different user before using it:

- pkill palemoon (you gotta close any existing instance, otherwise
  starting palemoon from another user will just connect to it)
- butch pack palemon (save a binary snapshot of old palemoon pkg)
- butch rebuild palemoon
- su otheruser # try it out
- if you encounter issues, you can now butch unpack palemoonxxxxx.tar
  from your backup and continue using your old profile.

this update gratuitously moves lots of files around in the file system,
many old mozbuild options were removed, and a new gratuitous check
for GCC versions < 7.1.0 introduced, even though it builds just fine
with the stock 6.5.0 gcc.
additionally once again we gotta download UXP (which was renamed to
GRE and trademarked) and palemoon separately from git tags, which
makes it more of a pain to deal to with it, and requires 300 MB of
downloads instead of 200 MB, due to gitea only offering gzip downloads,
not xz.