butch: add new command packapp
this allows to pack rarely used applications for smaller storage.
it requires zstd package installed.
the application will be butch-pack'd and put into a special dir,
wrappers for all executables will be created in /bin, that unpack
the package on-the-fly to /dev/shm and start it from there.

initial tests have shown space-saving up to 3x in case of qemu
(before: 198 MB, after: 50MB), and 1x in worst case (firefox).
thanks to zstd's high decompression speed, the startup overhead
is quite neglible, in case of qemu or firefox there's a delay
of like 1-2 seconds on my downclocked PC when starting up.

consider this applet to be in a beta-stage, i've already seen
that it will need improvement for some applications, we'll probably
need to keep files not in /bin and /lib available in the original
filesystem tree, or add some trickery to the build of eligible
applications so they access their files via relative paths
instead of absolute.