musl: bump to 1.2.3
it's finally done - musl 1.2.x made it into sabtoage.
i was holding back on this because musl 1.2.x introduces a new 64bit
time_t abi for 32bit archs, which has caused a lot of applications
to break in other distros (because of buggy assumption in those
apps) which have slowly been fixed upstream in the meantime.

additionally it has a new memory allocator which detects most forms
of UB and crashes the buggy application, which is nice but may break
some applications that worked by accident in the past too.

i've been testing 1.2.2 for several months on my x86_64 machine and
didn't notice any issues that haven't been fixed during the 1.2.3
release cycle, so it's finally time to make the switch.
it's expected that some packages will be bugged on 32 bit systems
so we'll have to wait for issue reports to pop up.