palemoon: adjust thread stack size to recently changed defaults
starting with 31.2.x, the thread stack size limits were changed[0],
giving way bigger stackframes to js stacks on several platforms
including linux; and since then i noticed regular crashes of the
browser; most likely due to the mismatch of what the browser
expected to have available vs what was actually requested from
the C library.

we now set these values to 2MB which is what the windows build
uses too, so it should be rather well tested.
the size is modified in palemoon sources and via linker argument
so musl can set the size to the expected value.

[0]: https://repo.palemoon.org/MoonchildProductions/UXP/issues/1805