ffmpeg: bump to 4.3.5, add libx265 and static binary support
libx265 support is optional, just like libx264 etc and will be picked up
if the library is installed.

if STATICBUILD=1 env var is passed, ffmpeg will be built as a static
executable but ALSO all the libraries will be available only as static
versions, which is probably not what you want, as those are quite big
and will be linked into every consumer.