add dosemu2 and related things
namely the fdpp distribution that was released at the same time.
even though the fdpp abomination* builds, it currently makes dosemu2
segfault as soon as it's loaded, because the bios elf file is somehow
corrupted. that's why its use is currently disabled in the dosemu2 recipe.
as soon as it is installed though, dosemu2 will pick it up even though it's
disabled, so beware.

* it truly _is_ an abomination. instead of just taking the freedos source
code as is, the author wrangles the C code through several layers of pre-
processing tools so he can compile it with a C++ compiler!
not only that, he uses some hacks only clang understands so everybody has
to use his favorite compiler, too.
really, as soon as somebody uses the term "plus plus", the alarm bells in
my head start to ring.
while what the author intends to do with the library/bios is probably not
that bad an idea, the way it is built is just mindbendingly fragile.
preprocessing C code through several shell utilities (GNU behaviour assumed)
and custom bison / flex parsers (naturally using the latest features available)
to compile it with a x86_64 C++ compiler, mixing it with 16 bit x86 asm
assembled by nasm, and linking it together with the system linker...
since it is fragile af, the author recommends to use pre-built binaries.
way to go.

fortunately, dosemu2 runs just fine without freedos "plus plus" as soon as
one copies the system files of a DOS installation into the c_drive directory.

anyway, i'm checking the defunct fdpp package into the repo because it was
quite an effort to get it to compile, and fixing it might not be that hard.
apart from that it's the right version that was released right together with
the dosemu2 release. it might even be possible to overwrite the bios elf file
with one distributed as binary, while keeping the library part from the musl-
linked package here.